Refactor footer component
