+++ src/app/(site)/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next';
 import dynamic from 'next/dynamic';
 import clsx from 'clsx';
 
-import Footer from '@/components/Footer';
+import Footer from '@/components/layout/Footer';
 import MainMenu from '@/components/layout/MainMenu';
 import MainLogo from '@/components/layout/MainLogo';
 import PathMarker from '@/components/layout/PathMarker';
