+++ src/styles/layout/footer.scss
@@ -1,0 +1,61 @@
+.site-footer {
+  padding-top: 1rem;
+  padding-bottom: 1rem;
+
+  @include tablet {
+    padding-top: 0.75rem;
+    padding-bottom: 0.75rem;
+  }
+
+  color: var(--color-white);
+  background-color: rgba(0, 0, 0, 0.67);
+
+  @media print {
+    display: none;
+  }
+
+  .container {
+    @include tablet {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+    }
+  }
+
+  .social {
+    display: flex;
+
+    @include mobile {
+      justify-content: space-evenly;
+      margin-bottom: 1rem;
+    }
+    @include tablet {
+      gap: 0.5rem;
+    }
+
+    a,
+    img {
+      display: block;
+      border-radius: 50%;
+    }
+  }
+
+  .copy {
+    text-align: center;
+    @include tablet {
+      text-align: right;
+    }
+
+    span {
+      display: block;
+
+      &:nth-child(1) {
+        font-size: 0.875rem;
+        text-transform: uppercase;
+      }
+      &:nth-child(2) {
+        font-size: 0.75rem;
+      }
+    }
+  }
+}
